requirements.txt
flask
flask-restplus
pymongo
PyYAML